Consolidate Platform enum and extract URL normalization utility
- Move Platform(str, Enum) from cli.py into config.py, replacing the
  PlatformType Literal alias -- single definition, used by both config
  models and CLI option type
- Use Platform.woocommerce / Platform.shopify in run_spider instead of
  raw string comparisons
- Extract normalize_url() to utils.py and use it in both get_sitemap_url
  and get_api_url, eliminating duplicated scheme-inference logic
